move config build to build/
