+++ packages/core/src/config/contracts.ts
@@ -1,7 +1,7 @@
 import type { Abi, Address } from "abitype";
 
-import type { ResolvedConfig } from "@/config/config";
 import type { Options } from "@/config/options";
+import type { ResolvedConfig } from "@/config/types";
 
 import { buildAbi } from "./abi";
 import { type Network, buildNetwork } from "./networks";
